Auth API implementation

+++ src/post/post.resolver.ts
@@ -1,30 +1,47 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
+import { UnauthorizedException } from '@nestjs/common';
 import { Args, ID, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { GetCurrentUser, Public } from '../common/decorators';
 import { PostCreateDto, PostGetDto } from './dto';
 import { PostService } from './post.service';
 import { Post } from './schema';
 
-@Resolver((of) => Post)
+@Resolver(() => Post)
 export class PostResolver {
     constructor(private postService: PostService) {}
 
-    @Mutation((returns) => Post)
-    createPost(@Args('createPostInput') dto: PostCreateDto): Promise<Post> {
+    @Mutation(() => Post)
+    createPost(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Args('createPostInput') dto: PostCreateDto
+    ): Promise<Post> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('unauthorized');
+        }
+
         return this.postService.create(dto);
     }
 
-    @Query((returns) => [Post])
-    posts(@Args({ name: 'getPostInput' }) dto: PostGetDto): Promise<Post[]> {
+    @Public()
+    @Query(() => [Post])
+    posts(@Args('getPostInput') dto: PostGetDto): Promise<Post[]> {
         return this.postService.findAll(dto);
     }
 
-    @Query((returns) => Post)
-    post(@Args({ name: 'id', type: () => ID }) postId: string): Promise<Post> {
+    @Public()
+    @Query(() => Post)
+    post(@Args('postId', { type: () => ID }) postId: string): Promise<Post> {
         return this.postService.findOne(postId);
     }
 
-    @Mutation((returns) => Post)
-    deletePost(@Args({ name: 'id', type: () => ID }) postId: string): Promise<Post> {
+    @Mutation(() => Post)
+    deletePost(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Args('postId', { type: () => ID }) postId: string
+    ): Promise<Post> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('unauthorized');
+        }
+
         return this.postService.delete(postId);
     }
 }
